chore: move 7 domain services from src/utils/ to src/services/
src/utils/ is now reserved for pure, stateless, domain-agnostic helpers.
Domain services with stateful lifecycle or heavy domain dependencies move to
src/services/ alongside the existing service layer:

  utils/ApplicationPropertiesFinder  → services/ApplicationPropertiesFinder
  utils/KameletFileReader             → services/KameletFileReader
  utils/MavenRuntimeDetector         → services/MavenRuntimeDetector
  utils/PortManager                   → services/PortManager
  utils/StepsOnSaveManager            → services/StepsOnSaveManager
  utils/SuggestionRegistry            → services/SuggestionRegistry
  utils/TestFolderResolver            → services/TestFolderResolver

Update all import sites across commands, executors, extension, services,
views, webview, and test files. Fix intra-service self-references
(../services/ → ./) in MavenRuntimeDetector and StepsOnSaveManager.
