+++ src/executors/api/CamelCommandAPI.ts
@@ -14,7 +14,7 @@ import {
 	RuntimeType,
 } from '../types/ExecutorTypes';
 import { CamelSettingsHelper } from '../helpers/CamelSettingsHelper';
-import { TestFolderResolver } from '../../utils/TestFolderResolver';
+import { TestFolderResolver } from '../../services/TestFolderResolver';
 import { arePathsEqual } from '../../utils/Path';
 import { KAOTO_EXECUTOR_INFRA_ARGUMENTS_SETTING_ID } from '../../constants';
 import { KaotoOutputChannel } from '../../extension/KaotoOutputChannel';
