+++ src/extension/ExtensionContextHandler.ts
@@ -88,14 +88,14 @@ import { NewCamelProjectCommand } from '../commands/NewCamelProjectCommand';
 import { CamelTaskFactory } from '../tasks/CamelTaskFactory';
 import { CamelCommandAPI } from '../executors/api/CamelCommandAPI';
 import { DeploymentsProvider } from '../views/deployments/DeploymentsProvider';
-import { PortManager } from '../utils/PortManager';
+import { PortManager } from '../services/PortManager';
 import { ParentItem } from '../views/deployments/ParentItem';
 import { RouteOperation } from '../types/RouteOperation';
 import { ChildItem } from '../views/deployments/ChildItem';
 import { RecommendationCore } from '@redhat-developer/vscode-extension-proposals';
 import { WhatsNewPanel } from './WhatsNewPanel';
 import { satisfies } from 'compare-versions';
-import { StepsOnSaveManager } from '../utils/StepsOnSaveManager';
+import { StepsOnSaveManager } from '../services/StepsOnSaveManager';
 import { Folder } from '../views/integrations/Folder';
 import { TestsProvider } from '../views/tests/TestsProvider';
 import { AbstractFolderTreeProvider } from '../views/shared/AbstractFolderTreeProvider';
