+++ src/services/SuggestionRegistry.ts
